Adds product text loader
